Renamed git_utils module
