+++ packages/opal-common/opal_common/git_utils/exceptions.py
