+++ packages/opal-common/opal_common/git_utils/tests/commit_viewer_test.py
@@ -19,7 +19,7 @@ from typing import List
 
 from git import Repo
 from git.objects import Commit
-from opal_common.git.commit_viewer import CommitViewer, VersionedNode
+from opal_common.git_utils.commit_viewer import CommitViewer, VersionedNode
 
 
 def node_paths(nodes: List[VersionedNode]) -> List[Path]:
